+++ Ork/Sniffs/Formatting/AlphabeticalMethodNamesSniff.php
@@ -63,7 +63,9 @@ class AlphabeticalMethodNamesSniff implements Sniff
         // We only care about class methods, so ignore global functions (level 0).
         if ($tokens[$stackPtr]['level'] === 1) {
             $methodName = $phpcsFile->getDeclarationName($stackPtr);
-            if ($this->lastMethodName !== null && $methodName <= $this->lastMethodName) {
+            if ($this->lastMethodName !== null
+                && strcasecmp($methodName, $this->lastMethodName) <= 0
+            ) {
                 $phpcsFile->addError('Method "%s" is not in alphabetical order.', $stackPtr, 'OutOfOrder', [$methodName]);
             }