+++ tests/Ork.Formatting.AlphabeticalMethodNames/ClassGood.php
@@ -1,0 +1,13 @@
+<?php
+
+class Foo
+{
+    protected function _a() { }
+    protected function _z() { }
+    protected function a() { }
+    protected function b() { }
+    protected function c() { }
+    protected function d() { }
+    protected function getBoxes() { }
+    protected function getBoxModel() { }
+}