+++ tests/Ork.Formatting.AlphabeticalMethodNames/SkipGlobals.php
@@ -1,0 +1,10 @@
+<?php
+
+function a() { }
+function _a() { }
+function b() { }
+function d() { }
+function c() { }
+function getBoxModel() { }
+function getBoxes() { }
+function _z() { }