+++ tests/Ork.Formatting.AlphabeticalPropertyNames/AbstractClassBad.php
@@ -1,0 +1,15 @@
+<?php
+
+abstract class AbstractFoo
+{
+    protected $a;
+    protected $_a;
+    protected $b;
+    public function bar() { }
+    protected $d;
+    public function foo() { }
+    protected $c;
+    protected $getBoxModel;
+    protected $getBoxes;
+    protected $_z
+}