+++ tests/Ork.Formatting.AlphabeticalPropertyNames/SkipArgsAndLocals.php
@@ -1,0 +1,10 @@
+<?php
+
+class Foo
+{
+    public function foo($b, $a)
+    {
+        $b = 0;
+        $a = 0;
+    }
+}